+++ crates/gen/src/delegate.rs
@@ -46,47 +46,7 @@ impl Delegate {
             self.name.gen_signature(&format!("{{{:#?}}}", &self.guid))
         };
 
-        let invoke_args = self
-            .method
-            .params
-            .iter()
-            .map(|param| param.gen_invoke_arg());
-
-        let invoke_upcall = if let Some(return_type) = &self.method.return_type {
-            if return_type.array {
-                let result = format_ident(&return_type.name);
-                let result_size = squote::format_ident!("array_size_{}", &return_type.name);
-
-                quote! {
-                    match ((*this).invoke)(#(#invoke_args,)*) {
-                        ::std::result::Result::Ok(ok__) => {
-                            let (ok_data__, ok_data_len__) = ok__.into_abi();
-                            *#result = ok_data__;
-                            *#result_size = ok_data_len__;
-                            ::winrt::ErrorCode(0)
-                        }
-                        ::std::result::Result::Err(err) => err.into()
-                    }
-                }
-            } else {
-                let return_name = format_ident(&return_type.name);
-
-                quote! {
-                    match ((*this).invoke)(#(#invoke_args,)*) {
-                        ::std::result::Result::Ok(ok__) => {
-                            *#return_name = ::std::mem::transmute_copy(&ok__);
-                            ::std::mem::forget(ok__);
-                            ::winrt::ErrorCode(0)
-                        }
-                        ::std::result::Result::Err(err) => err.into()
-                    }
-                }
-            }
-        } else {
-            quote! {
-                ((*this).invoke)(#(#invoke_args,)*).into()
-            }
-        };
+        let invoke_upcall = self.method.gen_upcall(quote! { ((*this).invoke) });
 
         quote! {
             #[repr(transparent)]
